chore: remove extra route and rename editor folder
