+++ backend/src/config/config.controller.ts
@@ -9,15 +9,10 @@ import {
 import { AuthGuard } from '../auth/auth.guard'
 import { ConfigService } from './config.service'
 
+@UseGuards(AuthGuard)
 @Controller('config')
 export class ConfigController {
   constructor(private readonly configService: ConfigService) {}
-
-  @Get('/tinymceApiKey')
-  @UseGuards(AuthGuard)
-  getTinymceApiKey() {
-    return this.configService.getTinymceApiKey()
-  }
 
   @Get('/:id')
   findById(@Param('id') id: string): string {
