+++ frontend/src/features/editor/components/LetterViewer.tsx
