+++ frontend/src/features/editor/components/TemplateEditor.tsx
